+++ monkey/monkey_island/cc/services/attack/technique_reports/__init__.py
@@ -9,7 +9,7 @@ from monkey_island.cc.models.attack.attack_mitigations import AttackMitigations
 from monkey_island.cc.services.attack.attack_config import AttackConfig
 from monkey_island.cc.services.config_schema.config_schema import SCHEMA
 from monkey_island.cc.services.config_schema.config_schema_per_attack_technique import (
-    get_config_schema_per_attack_technique,
+    ConfigSchemaPerAttackTechnique,
 )
 
 logger = logging.getLogger(__name__)
@@ -122,8 +122,8 @@ class AttackTechnique(object, metaclass=abc.ABCMeta):
             return disabled_msg
         if status == ScanStatus.UNSCANNED.value:
             if not cls.config_schema_per_attack_technique:
-                cls.config_schema_per_attack_technique = get_config_schema_per_attack_technique(
-                    SCHEMA
+                cls.config_schema_per_attack_technique = (
+                    ConfigSchemaPerAttackTechnique().get_config_schema_per_attack_technique(SCHEMA)
                 )
             unscanned_msg = cls._get_unscanned_msg_with_reasons(
                 cls.unscanned_msg, cls.config_schema_per_attack_technique
@@ -143,7 +143,7 @@ class AttackTechnique(object, metaclass=abc.ABCMeta):
             reasons.append(f"- Monkey did not run on any {cls.relevant_systems[0]} systems.")
         if cls.tech_id in config_schema_per_attack_technique:
             reasons.append(
-                "- The following configuration options were disabled:<br/>"
+                "- The following configuration options were disabled or empty:<br/>"
                 f"{cls._get_relevant_config_values(config_schema_per_attack_technique)}"
             )